builder(instance): wait for instance to start (#24)
Updated the plugin to wait for the instance to start right after the
instance is created. Normally this isn't an issue since the communicator
will wait to connect to the instance before running the provisioning
step but when the communicator is skipped the instance isn't given
enough time to start before we ask it to stop.