builder(instance): fix panic when build is interrupted (#31)
Fixed an issue where Packer would panic when the build is interrupted.
The panic occurred because the state bag was missing `image_id` and
`image_name` elements but the artifact construction was performing a
type assert on those elements unconditionally.

While in the code I noticed an issue where the instance create step
would fail to clean up a running instance if the build was interrupted
while the step was waiting for the instance to start. Similar to above,
this was because the `instance_id` and `boot_disk_id` elements were not
in the state bag and the `Cleanup` method conditionally cleans up the
resources based on those elements.